requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
faiss-cpu==1.8.0.post1
sentence-transformers==2.2.2
groq
numpy<2
